Digital wallet(people pay)
Description: Digital wallet(people pay)is an web application for providing transactions like online money sending and recharging the prepaid or post-paid cards to registered people pay users and recharge any prepaid cards.The idea of this project is to provide secure and faster services to the customer for the online transactions like sending money and recharge. The user will get benefit of doing transactions in a fraction of seconds without going to any nearer bank or stores for doing this transactions. The system to provide service with faster and efficient response, by their choice. This system focuses on providing transactions based on customer requirement. A separate interface for the user will facilitate interaction with expert system